Drop needless use of Cell, using () instead
By its design an Rc holds its own counter of references, which happens
to have in our case a value identical to the one we were maintaining
inside it with a Cell. Changing Rc<Cell> to Rc<()> seems to result in
approximately an 8% reduction in runtime for `mangle` without failing
any tests.